Add functions for converting extended public keys and getting key fingerprints
This commit adds two new functions, `convertPublicKey` and `getFingerprint`, which allow for the conversion of extended public keys to a desired prefix format and the retrieval of key fingerprints, respectively. These functions utilize the `bs58check` and `bip32` libraries to perform the necessary calculations and validations. The commit also includes the necessary data structures and constants for these functions to work, including a map of prefix formats and their associated version bytes, and a map of network types and their corresponding parameters.